toymakon-backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.3
Pillow>=10.0
python-dotenv>=1.0
django-filter>=24.0
PyMySQL>=1.1
